+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Any notable change to this project will be documented in this file.
 ## Unreleased
 
 - print display item count in console output
+- folder input support (e.g. popcorn -f dirname)
 
 ## [v0.0.1 Release (23-10-2023)](https://github.com/efleming-intel/popcorn/tree/v0.0.1)